requirements.txt
torch>=1.9.0
numpy>=1.19.0
matplotlib>=3.3.0
open3d>=0.13.0
build123d>=0.1.0
numpy-stl>=2.16.0
